requirements.txt
langgraph 
langsmith
torch
python-dotenv
groq